shotping_flask/requirements.txt
alembic==1.9.1
amqp==5.1.1
async-timeout==4.0.2
boto3==1.26.48
botocore==1.29.48
celery==5.2.7
cmake==3.25.0
decorator==5.1.1
elastic-transport==8.4.0
elasticsearch==7.17.8
Flask==2.2.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-restx==1.0.3
Flask-SQLAlchemy==3.0.2
gunicorn==20.1.0
gitpython>=3.1.30
ipython==8.8.0
matplotlib==3.6.2
numpy==1.24.0
opencv-python==4.7.0.68
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
prometheus-client==0.15.0
prometheus-flask-exporter==0.21.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
PyMySQL==1.0.2
PyYAML==6.0
psutil==5.9.4
redis==4.4.2
requests==2.28.1
s3transfer==0.6.0
setuptools>=65.5.1 
scipy==1.10.0
seaborn==0.12.1
six==1.16.0
SQLAlchemy==1.4.46
stack-data==0.6.2
torch==1.13.1
torchvision>=0.8.1
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0
ultralytics>=8.0.145
pytest===7.4.0